+++ src/config/camera.rs
@@ -1,20 +1,20 @@
 #[derive(Deserialize, Debug)]
 pub struct Camera {
-    pub fov: f64,
+    pub fov: f32,
     pub width: u32,
     pub height: u32,
-    pub position: [f64; 3],
-    pub look_at: [f64; 3],
-    pub up: [f64; 3],
+    pub position: [f32; 3],
+    pub look_at: [f32; 3],
+    pub up: [f32; 3],
 }
 
 impl Camera {
-    pub fn new(fov: f64,
+    pub fn new(fov: f32,
                width: u32,
                height: u32,
-               position: [f64; 3],
-               look_at: [f64; 3],
-               up: [f64; 3])
+               position: [f32; 3],
+               look_at: [f32; 3],
+               up: [f32; 3])
                -> Camera {
         Camera {
             fov: fov,
