+++ src/config/material.rs
@@ -2,11 +2,11 @@ use material::IllumninationModel;
 
 #[derive(Deserialize, Debug, Copy, Clone)]
 pub struct Material {
-    pub ambient_color: [f64; 3],
-    pub diffuse_color: [f64; 3],
-    pub specular_color: [f64; 3],
-    pub specular_exponent: f64,
+    pub ambient_color: [f32; 3],
+    pub diffuse_color: [f32; 3],
+    pub specular_color: [f32; 3],
+    pub specular_exponent: f32,
     pub illumination_model: IllumninationModel,
-    pub reflection_coefficient: Option<f64>,
-    pub refraction_coefficient: Option<f64>,
+    pub reflection_coefficient: Option<f32>,
+    pub refraction_coefficient: Option<f32>,
 }
